+++ app/javascript/components/help/FireSection.tsx
@@ -3,15 +3,14 @@ import { helpLink } from "./helpData";
 import { SectionProps } from "../game/HelpDisplay";
 
 export default function FireSection({ section }: SectionProps) {
-
   return (
     <div>
       <p>
         Firing is the act of one (or more) units attacking one (or more) opponent units, possibly
         breaking, reducing, or eliminating them. The rules for firing,{" "}
         {helpLink("Intensive Fire", "intensive fire")}, and{" "}
-        {helpLink("Reaction Fire", "reaction fire")} are the same, except a few different
-        modifiers apply.
+        {helpLink("Reaction Fire", "reaction fire")} are the same, except a few different modifiers
+        apply.
       </p>
       <h3>{section}.1. Firing</h3>
       <p>
@@ -19,9 +18,10 @@ export default function FireSection({ section }: SectionProps) {
         Pinned and tired units may fire, albeit at a penalty. Weapons operated by pinned or broken
         or exhausted units may not fire — if you plan to use intensive fire separately with a
         machine gun and the unit carrying it, fire the machine gun first. Crewed weapons may not
-        fire if the operator is activated. Crews manning crewed weapons may not fire (the crewed weapon
-        must be fired instead, until and unless the crew unmans the weapon). Crews will be activated
-        when the weapon is fired.
+        fire if the operator is activated. Crews manning crewed weapons may not fire (the crewed
+        weapon must be fired instead, until and unless the crew unmans the weapon). Crews will be
+        activated when the weapon is fired. Units being carried or towed by vehicles may not fire{" "}
+        <strong>except</strong> radios being carried by infantry may call in artillery.
       </p>
       <h3>{section}.2. Targeting Infantry Weapons and Crewed Weapons</h3>
       <p>
@@ -40,8 +40,8 @@ export default function FireSection({ section }: SectionProps) {
         firing units is used if they aren&apos;t all in the same hex. If all units in the fire group
         have rapid fire (except the leader), a rapid fire attack may be performed (see below),
         otherwise an infantry attack will be performed. To select a fire group in the game, select a
-        unit in the same hex as the leader (it shouldn&apos;t be the leader, but the leader
-        must be the next unit selected to add additional units).
+        unit in the same hex as the leader (it shouldn&apos;t be the leader, but the leader must be
+        the next unit selected to add additional units).
       </p>
       <p>
         The only exception to units firing individually is that a carried weapon (that doesn&apos;t